requirements.txt
Django==2.0.4
djangorestframework
pygments
psycopg2
mysqlclient